requirements.txt
mmcv-full==1.5.0
torch~=2.0.1
numpy
requests
scipy~=1.11.1
torchvision~=0.15.2
decord ~= 0.6.0
